fix: Remove deleted articles that failed to sync
These articles were deleted within the GitBook dashboard but it hasn't
been reflected in this repository probably due to a bug in GitBook's
synchronization feature.

Delete these articles manually to unblock release of Suite.

Do not delete buy-sell-and-trade.md and passphrase.md pages in the root
of the repo. These don't get picked up by Suite as they're not put in a
locale and they might come handy for the GitBook folks when debugging
the synchronization issue.